requirements.txt
Django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
numpy==1.26.4
scipy==1.13.0
pandas==2.2.1
scikit-learn==1.5.2
yfinance==0.2.45
tensorflow==2.15.0 